Move Configure sections to docs
