+++ packages/demo/frontend/src/components/docs/ConfigureSignersSection.tsx
@@ -3,23 +3,25 @@ import { colors } from '@/constants/colors'
 import PrivyLogo from '@/assets/privy-logo-white.svg'
 import DynamicLogo from '@/assets/dynamic-logo-white.svg'
 import TurnkeyLogo from '@/assets/turnkey-logo-white.svg'
-import TabbedCodeBlock from './TabbedCodeBlock'
+import TabbedCodeBlock from '../home/TabbedCodeBlock'
 
-interface ConfigureWalletsSectionProps {
+interface ConfigureSignersSectionProps {
   stepNumber: number
   isOpen: boolean
   onToggle: () => void
 }
 
-function ConfigureWalletsSection({
+function ConfigureSignersSection({
   stepNumber,
   isOpen,
   onToggle,
-}: ConfigureWalletsSectionProps) {
+}: ConfigureSignersSectionProps) {
   const [selectedWalletProvider, setSelectedWalletProvider] = useState('privy')
-  const [selectedPrivyTab, setSelectedPrivyTab] = useState('frontend')
-  const [selectedDynamicTab, setSelectedDynamicTab] = useState('frontend')
-  const [selectedTurnkeyTab, setSelectedTurnkeyTab] = useState('frontend')
+  const [selectedSmartPrivyTab, setSelectedSmartPrivyTab] = useState('frontend')
+  const [selectedSmartDynamicTab, setSelectedSmartDynamicTab] =
+    useState('frontend')
+  const [selectedSmartTurnkeyTab, setSelectedSmartTurnkeyTab] =
+    useState('frontend')
 
   const privyFrontendCode = `import { actions } from './config'
 import { useWallets } from '@privy-io/react-auth'
@@ -30,58 +32,83 @@ const embeddedWallet = wallets.find(
   (wallet) => wallet.walletClientType === 'privy',
 )
 
-// ACTIONS: Let wallet make onchain Actions
-const wallet = await actions.wallet.hostedWalletToActionsWallet({
+// ACTIONS: Create signer from hosted wallet
+const signer = await actions.wallet.createSigner({
   connectedWallet: embeddedWallet,
+})
+
+// ACTIONS: Create smart wallet capable of Actions
+const { wallet } = await actions.wallet.createSmartWallet({
+  signer: signer
 })`
 
   const privyBackendCode = `import { actions } from './config'
 import { PrivyClient } from '@privy-io/node'
-
-const privyClient = new PrivyClient(env.PRIVY_APP_ID, env.PRIVY_APP_SECRET)
+import { getAddress } from 'viem'
 
 // PRIVY: Create wallet
 const privyWallet = await privyClient.walletApi.createWallet({
   chainType: 'ethereum',
 })
 
-// ACTIONS: Let wallet make onchain Actions
-const wallet = await actions.wallet.hostedWalletToActionsWallet({
+// ACTIONS: Create signer from hosted wallet
+const privySigner = await actions.wallet.createSigner({
   walletId: privyWallet.id,
-  address: privyWallet.address,
+  address: getAddress(privyWallet.address),
+})
+
+// ACTIONS: Create smart wallet capable of Actions
+const { wallet } = await actions.wallet.createSmartWallet({
+  signer: privySigner
 })`
 
-  const dynamicCode = `import { actions } from './config'
+  const dynamicFrontendCode = `import { actions } from './config'
 import { useDynamicContext } from "@dynamic-labs/sdk-react-core"
 
 // DYNAMIC: Fetch wallet
 const { primaryWallet } = useDynamicContext()
-const embeddedWallet = primaryWallet
 
-// ACTIONS: Let wallet make onchain Actions
-const wallet = await actions.wallet.hostedWalletToVerbsWallet({
-  wallet: embeddedWallet,
+// ACTIONS: Create signer from hosted wallet
+const signer = await actions.wallet.createSigner({wallet: primaryWallet})
+
+// ACTIONS: Create smart wallet capable of Actions
+const { wallet } = await actions.wallet.createSmartWallet({
+  signer: signer
 })`
+
+  const dynamicBackendCode = ``
 
   const turnkeyFrontendCode = `import { actions } from './config'
 import { useTurnkey } from "@turnkey/react-wallet-kit"
 
 // TURNKEY: Fetch wallet
-const { wallets, createWallet, refreshWallets, httpClient, session } = useTurnkey()
+const { wallets, user, createWallet, refreshWallets, httpClient, session } = useTurnkey()
+useEffect(() => {
+  async function createEmbeddedWallet() {
+    const wallet = await createWallet({
+      walletName: \`My New Wallet \${Math.random().toString(36).substring(2, 15)}\`,
+      accounts: ["ADDRESS_FORMAT_ETHEREUM"],
+    })
+    refreshWallets()
+  }
 
-const embeddedWallet = await createWallet({
-  walletName: \`My New Wallet \${Math.random()}\`,
-  accounts: ["ADDRESS_FORMAT_ETHEREUM"],
-})
+const embeddedWallet = wallets.find(
+  (wallet) => wallet.accounts.some((account) => account.addressFormat === 'ADDRESS_FORMAT_ETHEREUM' && wallet.source === WalletSource.Embedded,
+)
 
 const walletAddress = embeddedWallet.accounts[0].address
 
-// ACTIONS: Let wallet make onchain Actions
-const wallet = await actions.wallet.hostedWalletToActionsWallet({
+// ACTIONS: Create signer from hosted wallet
+const signer = await actions.wallet.createSigner({
   client: httpClient,
   organizationId: session.organizationId,
   signWith: walletAddress,
-  ethereumAddress: walletAddress,
+  ethereumAddress: walletAddress
+})
+
+// ACTIONS: Create smart wallet capable of Actions
+const { wallet } = await actions.wallet.createSmartWallet({
+  signer: signer
 })`
 
   const turnkeyBackendCode = `import { actions } from './config'
@@ -105,10 +132,15 @@ const turnkeyWallet = await turnkeyClient.apiClient().createWallet({
   }],
 })
 
-// ACTIONS: Let wallet make onchain Actions
-const wallet = await actions.wallet.hostedWalletToActionsWallet({
+// ACTIONS: Create signer from hosted wallet
+const turnkeySigner = await actions.wallet.createSigner({
   organizationId: turnkeyWallet.activity.organizationId,
   signWith: turnkeyWallet.addresses[0],
+})
+
+// ACTIONS: Create smart wallet capable of Actions
+const { wallet } = await actions.wallet.createSmartWallet({
+  signer: turnkeySigner
 })`
 
   return (
@@ -133,7 +165,7 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
             className="text-lg font-medium"
             style={{ color: colors.text.cream }}
           >
-            Configure Wallets
+            Configure Signers
           </h3>
         </div>
         <svg
@@ -162,7 +194,7 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
       >
         <div className="pt-6 pb-4">
           <p className="text-base mb-4" style={{ color: colors.text.cream }}>
-            Actions supports your existing embedded wallet provider.
+            Use embedded wallets as signers of smart wallets you control.
           </p>
           <div
             className="rounded-lg overflow-hidden mb-8 shadow-2xl"
@@ -172,13 +204,9 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
                 '0 25px 50px -12px rgba(0, 0, 0, 0.5), 0 0 20px rgba(184, 187, 38, 0.05)',
             }}
           >
-            {/* Tab switcher with logos */}
             <div
               className="flex border-b"
-              style={{
-                backgroundColor: colors.bg.code,
-                borderColor: 'rgba(184, 187, 38, 0.15)',
-              }}
+              style={{ borderColor: 'rgba(184, 187, 38, 0.15)' }}
             >
               <button
                 onClick={() => setSelectedWalletProvider('privy')}
@@ -265,23 +293,25 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
                   </div>
 
                   <div>
-                    <p className="mb-2" style={{ color: colors.text.cream }}>
-                      2. Create a frontend or backend user wallet and extend it
-                      with DeFi Actions:
+                    <p
+                      className="text-base mb-2"
+                      style={{ color: colors.text.cream }}
+                    >
+                      2. Hosted user wallets can become signers for new,
+                      customizable smart wallets:
                     </p>
                     <TabbedCodeBlock
                       tabs={[
                         { label: 'Frontend', code: privyFrontendCode },
                         { label: 'Backend', code: privyBackendCode },
                       ]}
-                      selectedTab={selectedPrivyTab}
-                      onTabChange={setSelectedPrivyTab}
+                      selectedTab={selectedSmartPrivyTab}
+                      onTabChange={setSelectedSmartPrivyTab}
                       filename="wallet.ts"
                     />
                   </div>
                 </div>
               )}
-
               {selectedWalletProvider === 'dynamic' && (
                 <div className="space-y-6">
                   <div>
@@ -291,7 +321,7 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
                     >
                       1.{' '}
                       <a
-                        href="https://www.dynamic.xyz/docs/wallets/embedded-wallets/mpc/setup"
+                        href="https://docs.dynamic.xyz/quickstart"
                         target="_blank"
                         rel="noopener noreferrer"
                         className="text-blue-400 hover:text-blue-300 underline"
@@ -303,23 +333,25 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
                   </div>
 
                   <div>
-                    <p className="mb-2" style={{ color: colors.text.cream }}>
-                      2. Create a frontend user wallet and extend it with DeFi
-                      Actions:
+                    <p
+                      className="text-base mb-2"
+                      style={{ color: colors.text.cream }}
+                    >
+                      2. Hosted user wallets can become signers for new,
+                      customizable smart wallets:
                     </p>
                     <TabbedCodeBlock
                       tabs={[
-                        { label: 'Frontend', code: dynamicCode },
-                        { label: 'Backend', code: '' },
+                        { label: 'Frontend', code: dynamicFrontendCode },
+                        { label: 'Backend', code: dynamicBackendCode },
                       ]}
-                      selectedTab={selectedDynamicTab}
-                      onTabChange={setSelectedDynamicTab}
+                      selectedTab={selectedSmartDynamicTab}
+                      onTabChange={setSelectedSmartDynamicTab}
                       filename="wallet.ts"
                     />
                   </div>
                 </div>
               )}
-
               {selectedWalletProvider === 'turnkey' && (
                 <div className="space-y-6">
                   <div>
@@ -341,17 +373,20 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
                   </div>
 
                   <div>
-                    <p className="mb-2" style={{ color: colors.text.cream }}>
-                      2. Create a frontend or backend user wallet and extend it
-                      with DeFi Actions:
+                    <p
+                      className="text-base mb-2"
+                      style={{ color: colors.text.cream }}
+                    >
+                      2. Hosted user wallets can become signers for new,
+                      customizable smart wallets:
                     </p>
                     <TabbedCodeBlock
                       tabs={[
                         { label: 'Frontend', code: turnkeyFrontendCode },
                         { label: 'Backend', code: turnkeyBackendCode },
                       ]}
-                      selectedTab={selectedTurnkeyTab}
-                      onTabChange={setSelectedTurnkeyTab}
+                      selectedTab={selectedSmartTurnkeyTab}
+                      onTabChange={setSelectedSmartTurnkeyTab}
                       filename="wallet.ts"
                     />
                   </div>
@@ -365,4 +400,4 @@ const wallet = await actions.wallet.hostedWalletToActionsWallet({
   )
 }
 
-export default ConfigureWalletsSection
+export default ConfigureSignersSection
